reorganize renderer, impl logic blending approximations, fix depth texture
