+++ crates/lazuli/src/system/gx/tev.rs
@@ -9,7 +9,7 @@ use bitos::integer::u3;
 
 #[bitos(3)]
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum Channel {
+pub enum Input {
     Channel0            = 0x0,
     Channel1            = 0x1,
     Reserved0           = 0x2,
@@ -30,7 +30,7 @@ pub struct StageRefs {
     #[bits(6)]
     pub map_enable: bool,
     #[bits(7..10)]
-    pub channel: Channel,
+    pub input: Input,
 }
 
 #[bitos(32)]
