+++ crates/renderer/src/render.rs
@@ -1,37 +1,25 @@
 mod data;
 mod framebuffer;
 mod pipeline;
-mod sampler;
 mod texture;
 
 use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::{Arc, Mutex};
 
 use glam::{Mat4, Vec2};
-use lazuli::modules::render::oneshot::Sender;
-use lazuli::modules::render::{
-    Action, ClutData, ClutId, ClutRef, CopyArgs, Sampler, Scaling, TexEnvConfig, TexGenConfig,
-    Texels, Texture, TextureId, Viewport, XfbPart, oneshot,
-};
+use lazuli::modules::render::{Action, Sampler, Scaling, Viewport};
 use lazuli::system::gx::color::Rgba;
-use lazuli::system::gx::pix::{
-    self, BlendMode, ColorCopyFormat, CompareMode, ConstantAlpha, DepthCopyFormat, DepthMode,
-    DstBlendFactor, Scissor, SrcBlendFactor,
-};
-use lazuli::system::gx::xform::{ChannelControl, Light};
-use lazuli::system::gx::{
-    CullingMode, EFB_HEIGHT, EFB_WIDTH, MatrixId, Topology, Vertex, VertexStream, tev,
-};
-use lazuli::system::vi::Dimensions;
+use lazuli::system::gx::pix::{ConstantAlpha, Scissor};
+use lazuli::system::gx::xform::{Channel, Light};
+use lazuli::system::gx::{EFB_HEIGHT, EFB_WIDTH, MatrixId, Topology, Vertex, VertexStream};
 use rustc_hash::FxBuildHasher;
 use schnellru::{ByLength, LruMap};
 use seq_macro::seq;
-use zerocopy::{FromBytes, IntoBytes};
+use zerocopy::IntoBytes;
 
 use crate::alloc::Allocator;
 use crate::blit::{ColorBlitter, Converter, DepthBlitter};
 use crate::clear::Cleaner;
-use crate::render::pipeline::TexGenStageSettings;
 use crate::render::texture::TextureRef;
 
 pub struct Shared {
@@ -90,7 +78,6 @@ pub struct Renderer {
     // caches
     pipeline_cache: pipeline::Cache,
     texture_cache: texture::Cache,
-    sampler_cache: sampler::Cache,
     textures_group_cache: GroupCache<TexturesGroupEntries>,
 
     // state
@@ -109,19 +96,6 @@ pub struct Renderer {
     actions: u64,
 }
 
-fn set_channel(channel: &mut data::Channel, control: ChannelControl) {
-    channel.material_from_vertex = control.material_from_vertex() as u32;
-    channel.ambient_from_vertex = control.ambient_from_vertex() as u32;
-    channel.lighting_enabled = control.lighting_enabled() as u32;
-    channel.diffuse_attenuation = control.diffuse_attenuation() as u32;
-    channel.attenuation = control.attenuation() as u32;
-    channel.specular = !control.not_specular() as u32;
-
-    let a = control.lights0to3();
-    let b = control.lights4to7();
-    channel.light_mask = [a[0], a[1], a[2], a[3], b[0], b[1], b[2], b[3]].map(|b| b as u32);
-}
-
 impl Renderer {
     pub fn new(device: wgpu::Device, queue: wgpu::Queue) -> (Self, Arc<Shared>) {
         let embedded_fb = framebuffer::Embedded::new(&device);
@@ -134,7 +108,6 @@ impl Renderer {
 
         let pipeline_cache = pipeline::Cache::new(&device);
         let texture_cache = texture::Cache::default();
-        let sampler_cache = sampler::Cache::default();
 
         let color = embedded_fb.color();
         let multisampled_color = embedded_fb.multisampled_color();
@@ -206,7 +179,6 @@ impl Renderer {
 
             pipeline_cache,
             texture_cache,
-            sampler_cache,
             textures_group_cache: LruMap::with_hasher(ByLength::new(512), FxBuildHasher),
 
             viewport: Default::default(),
@@ -231,12 +203,12 @@ impl Renderer {
     pub fn exec(&mut self, action: Action) {
         match action {
             Action::SetXfbDimensions(dims) => self.set_xfb_dimensions(dims),
-            Action::SetFramebufferFormat(fmt) => self.set_framebuffer_format(fmt),
+            Action::SetEfbFormat(fmt) => self.set_efb_format(fmt),
             Action::SetViewport(viewport) => self.set_viewport(viewport),
             Action::SetScissor(scissor) => self.set_scissor(scissor),
             Action::SetCullingMode(mode) => self.set_culling_mode(mode),
             Action::SetClearColor(color) => self.set_clear_color(color),
-            Action::SetClearDepth(depth) => self.clear_depth = depth,
+            Action::SetClearDepth(depth) => self.set_clear_depth(depth),
             Action::SetBlendMode(mode) => self.set_blend_mode(mode),
             Action::SetDepthMode(mode) => self.set_depth_mode(mode),
             Action::SetAlphaFunction(func) => self.set_alpha_function(func),
@@ -324,36 +296,6 @@ impl Renderer {
         idx as u32
     }
 
-    fn set_xfb_dimensions(&mut self, dims: Dimensions) {
-        if dims == self.external_fb.dimensions() {
-            return;
-        }
-
-        self.external_fb.resize(
-            &self.device,
-            wgpu::Extent3d {
-                width: dims.width as u32,
-                height: dims.height as u32,
-                depth_or_array_layers: 1,
-            },
-        );
-
-        let mut output = self.shared.output.lock().unwrap();
-        *output = self.external_fb.framebuffer().clone();
-    }
-
-    fn set_framebuffer_format(&mut self, format: pix::BufferFormat) {
-        self.flush(format_args!("framebuffer format changed to {format:?}"));
-
-        match format {
-            pix::BufferFormat::RGB8Z24 | pix::BufferFormat::RGB565Z16 => {
-                self.pipeline_settings.has_alpha = false
-            }
-            pix::BufferFormat::RGBA6Z24 => self.pipeline_settings.has_alpha = true,
-            _ => (),
-        }
-    }
-
     fn apply_scissor_and_viewport(&mut self) {
         let (scissor_x, scissor_y) = self.scissor.top_left();
         let (scissor_width, scissor_height) = self.scissor.dimensions();
@@ -401,103 +343,12 @@ impl Renderer {
         }
     }
 
-    fn set_culling_mode(&mut self, mode: CullingMode) {
-        if self.pipeline_settings.culling != mode {
-            self.flush(format_args!("changed culling mode to {mode:?}"));
-            self.pipeline_settings.culling = mode;
-        }
-    }
-
     fn set_clear_color(&mut self, rgba: Rgba) {
-        self.debug(format!("set clear color to {rgba:?}"));
         self.clear_color = rgba;
     }
 
-    fn set_blend_mode(&mut self, mode: BlendMode) {
-        let (src, dst, op) = if mode.blend_subtract() {
-            (
-                wgpu::BlendFactor::One,
-                wgpu::BlendFactor::One,
-                wgpu::BlendOperation::ReverseSubtract,
-            )
-        } else {
-            let src = match mode.src_factor() {
-                SrcBlendFactor::Zero => wgpu::BlendFactor::Zero,
-                SrcBlendFactor::One => wgpu::BlendFactor::One,
-                SrcBlendFactor::DstColor => wgpu::BlendFactor::Dst,
-                SrcBlendFactor::InverseDstColor => wgpu::BlendFactor::OneMinusDst,
-                SrcBlendFactor::SrcAlpha => wgpu::BlendFactor::Src1Alpha,
-                SrcBlendFactor::InverseSrcAlpha => wgpu::BlendFactor::OneMinusSrc1Alpha,
-                SrcBlendFactor::DstAlpha => wgpu::BlendFactor::DstAlpha,
-                SrcBlendFactor::InverseDstAlpha => wgpu::BlendFactor::OneMinusDstAlpha,
-            };
-
-            let dst = match mode.dst_factor() {
-                DstBlendFactor::Zero => wgpu::BlendFactor::Zero,
-                DstBlendFactor::One => wgpu::BlendFactor::One,
-                DstBlendFactor::SrcColor => wgpu::BlendFactor::Src1,
-                DstBlendFactor::InverseSrcColor => wgpu::BlendFactor::OneMinusSrc1,
-                DstBlendFactor::SrcAlpha => wgpu::BlendFactor::Src1Alpha,
-                DstBlendFactor::InverseSrcAlpha => wgpu::BlendFactor::OneMinusSrc1Alpha,
-                DstBlendFactor::DstAlpha => wgpu::BlendFactor::DstAlpha,
-                DstBlendFactor::InverseDstAlpha => wgpu::BlendFactor::OneMinusDstAlpha,
-            };
-
-            (src, dst, wgpu::BlendOperation::Add)
-        };
-
-        let blend = pipeline::BlendSettings {
-            enabled: mode.enable(),
-            src,
-            dst,
-            op,
-            color_write: mode.color_mask(),
-            alpha_write: mode.alpha_mask(),
-        };
-
-        if self.pipeline_settings.blend != blend {
-            self.flush(format_args!("set blend settings to {blend:?}"));
-            self.pipeline_settings.blend = blend;
-        }
-    }
-
-    fn set_depth_mode(&mut self, mode: DepthMode) {
-        let compare = match mode.compare() {
-            CompareMode::Never => wgpu::CompareFunction::Never,
-            CompareMode::Less => wgpu::CompareFunction::Less,
-            CompareMode::Equal => wgpu::CompareFunction::Equal,
-            CompareMode::LessOrEqual => wgpu::CompareFunction::LessEqual,
-            CompareMode::Greater => wgpu::CompareFunction::Greater,
-            CompareMode::NotEqual => wgpu::CompareFunction::NotEqual,
-            CompareMode::GreaterOrEqual => wgpu::CompareFunction::GreaterEqual,
-            CompareMode::Always => wgpu::CompareFunction::Always,
-        };
-
-        let depth = pipeline::DepthSettings {
-            enabled: mode.enable(),
-            write: mode.update(),
-            compare,
-        };
-
-        if self.pipeline_settings.depth != depth {
-            self.flush(format_args!("set depth settings to {depth:?}"));
-            self.pipeline_settings.depth = depth;
-        }
-    }
-
-    fn set_alpha_function(&mut self, func: tev::alpha::Function) {
-        let settings = pipeline::AlphaFuncSettings {
-            comparison: func.comparison(),
-            logic: func.logic(),
-        };
-
-        if self.pipeline_settings.shader.texenv.alpha_func != settings {
-            self.flush(format_args!("set alpha function to {func:?}"));
-            self.pipeline_settings.shader.texenv.alpha_func = settings;
-        }
-
-        self.current_config.alpha_refs = func.refs().map(|x| x as u32);
-        self.current_config_dirty = true;
+    fn set_clear_depth(&mut self, depth: f32) {
+        self.clear_depth = depth;
     }
 
     fn set_constant_alpha_mode(&mut self, mode: ConstantAlpha) {
@@ -520,100 +371,24 @@ impl Renderer {
         self.current_config_dirty = true;
     }
 
-    fn set_color_channel(&mut self, idx: u8, control: ChannelControl) {
-        set_channel(
-            &mut self.current_config.color_channels[idx as usize],
-            control,
-        );
+    fn set_color_channel(&mut self, idx: u8, control: Channel) {
+        self.current_config.color_channels[idx as usize].update(control);
         self.current_config_dirty = true;
     }
 
-    fn set_alpha_channel(&mut self, idx: u8, control: ChannelControl) {
-        set_channel(
-            &mut self.current_config.alpha_channels[idx as usize],
-            control,
-        );
+    fn set_alpha_channel(&mut self, idx: u8, control: Channel) {
+        self.current_config.alpha_channels[idx as usize].update(control);
         self.current_config_dirty = true;
     }
 
     fn set_light(&mut self, idx: u8, light: Light) {
-        let data = &mut self.current_config.lights[idx as usize];
-        data.color = light.color.into();
-        data.cos_attenuation = light.cos_attenuation;
-        data.dist_attenuation = light.dist_attenuation;
-        data.position = light.position;
-        data.direction = light.direction;
+        self.current_config.lights[idx as usize].update(light);
         self.current_config_dirty = true;
     }
 
     fn set_projection_mat(&mut self, mat: Mat4) {
         self.current_config.projection_mat = mat;
         self.current_config_dirty = true;
-    }
-
-    fn set_texenv_config(&mut self, config: TexEnvConfig) {
-        self.flush(format_args!("texenv changed"));
-        self.pipeline_settings
-            .shader
-            .texenv
-            .stages
-            .clone_from(&config.stages);
-        self.pipeline_settings.shader.texenv.depth_tex = config.depth_tex;
-        self.current_config.consts = config.constants.map(Rgba::from);
-        self.current_config_dirty = true;
-    }
-
-    fn set_texgen_config(&mut self, config: TexGenConfig) {
-        self.flush(format_args!("texgen changed"));
-        self.pipeline_settings.shader.texgen.stages = config
-            .stages
-            .iter()
-            .map(|s| TexGenStageSettings {
-                base: s.base.clone(),
-                normalize: s.normalize,
-            })
-            .collect();
-
-        for (setting, value) in self
-            .current_config
-            .post_transform_mat
-            .iter_mut()
-            .zip(config.stages.iter().map(|s| s.post_matrix))
-        {
-            *setting = value;
-        }
-
-        self.current_config_dirty = true;
-    }
-
-    fn load_texture(&mut self, id: TextureId, texture: Texture) {
-        self.texture_cache.update_raw(id, texture);
-    }
-
-    fn load_clut(&mut self, id: ClutId, clut: ClutData) {
-        self.texture_cache.update_clut(id, clut);
-    }
-
-    fn set_texture_slot(
-        &mut self,
-        slot: usize,
-        id: TextureId,
-        clut: ClutRef,
-        sampler: Sampler,
-        scaling: Scaling,
-    ) {
-        let settings = TexSlotSettings {
-            texture: TextureRef { id, clut },
-            sampler,
-            scaling,
-        };
-
-        if self.tex_slots[slot] == settings {
-            return;
-        }
-
-        self.flush(format_args!("texture slot changed"));
-        self.tex_slots[slot] = settings;
     }
 
     fn flush_config(&mut self) {
@@ -790,6 +565,7 @@ impl Renderer {
             .clone()
     }
 
+    /// Flushes all pending draws as a single draw call.
     fn flush(&mut self, reason: std::fmt::Arguments) {
         if self.vertices.is_empty() {
             return;
@@ -828,13 +604,15 @@ impl Renderer {
 
         let textures = self.tex_slots.map(|s| {
             self.texture_cache
-                .get(&self.device, &self.queue, s.texture)
+                .get_texture(&self.device, &self.queue, s.texture)
                 .clone()
         });
 
-        let samplers = self
-            .tex_slots
-            .map(|s| self.sampler_cache.get(&self.device, s.sampler).clone());
+        let samplers = self.tex_slots.map(|s| {
+            self.texture_cache
+                .get_sampler(&self.device, s.sampler)
+                .clone()
+        });
 
         let textures_group = self.get_textures_group(TexturesGroupEntries { textures, samplers });
 
@@ -922,409 +700,5 @@ impl Renderer {
         self.textures_group_cache.clear();
 
         self.shared.rendered_anything.store(true, Ordering::Relaxed);
-    }
-
-    fn copy_color_to_tex(
-        &mut self,
-        x: u16,
-        y: u16,
-        width: u16,
-        height: u16,
-        half: bool,
-        encoder: &mut wgpu::CommandEncoder,
-    ) -> wgpu::TextureView {
-        let divisor = if half { 2 } else { 1 };
-        let size = wgpu::Extent3d {
-            width: width as u32 / divisor,
-            height: height as u32 / divisor,
-            depth_or_array_layers: 1,
-        };
-
-        let texture = self.device.create_texture(&wgpu::TextureDescriptor {
-            label: None,
-            dimension: wgpu::TextureDimension::D2,
-            size,
-            format: wgpu::TextureFormat::Rgba8Unorm,
-            usage: wgpu::TextureUsages::TEXTURE_BINDING
-                | wgpu::TextureUsages::RENDER_ATTACHMENT
-                | wgpu::TextureUsages::COPY_SRC,
-            view_formats: &[],
-            mip_level_count: 1,
-            sample_count: 1,
-        });
-        let view = texture.create_view(&Default::default());
-
-        let color = self.embedded_fb.color();
-        self.color_blitter.blit_to_texture(
-            &self.device,
-            color,
-            wgpu::Origin3d {
-                x: x as u32,
-                y: y as u32,
-                z: 0,
-            },
-            wgpu::Extent3d {
-                width: width as u32,
-                height: height as u32,
-                depth_or_array_layers: 1,
-            },
-            &view,
-            encoder,
-        );
-
-        view
-    }
-
-    fn copy_depth_to_tex(
-        &mut self,
-        x: u16,
-        y: u16,
-        width: u16,
-        height: u16,
-        half: bool,
-        encoder: &mut wgpu::CommandEncoder,
-    ) -> wgpu::TextureView {
-        let divisor = if half { 2 } else { 1 };
-        let size = wgpu::Extent3d {
-            width: width as u32 / divisor,
-            height: height as u32 / divisor,
-            depth_or_array_layers: 1,
-        };
-
-        let texture = self.device.create_texture(&wgpu::TextureDescriptor {
-            label: None,
-            dimension: wgpu::TextureDimension::D2,
-            size,
-            format: wgpu::TextureFormat::R32Float,
-            usage: wgpu::TextureUsages::TEXTURE_BINDING
-                | wgpu::TextureUsages::RENDER_ATTACHMENT
-                | wgpu::TextureUsages::COPY_SRC,
-            view_formats: &[],
-            mip_level_count: 1,
-            sample_count: 1,
-        });
-        let view = texture.create_view(&Default::default());
-
-        let depth = self.embedded_fb.depth();
-        self.depth_blitter.blit_to_texture(
-            &self.device,
-            depth,
-            wgpu::Origin3d {
-                x: x as u32,
-                y: y as u32,
-                z: 0,
-            },
-            wgpu::Extent3d {
-                width: width as u32,
-                height: height as u32,
-                depth_or_array_layers: 1,
-            },
-            &view,
-            encoder,
-        );
-
-        view
-    }
-
-    fn encode_color(
-        &mut self,
-        color: &wgpu::TextureView,
-        format: ColorCopyFormat,
-        encoder: &mut wgpu::CommandEncoder,
-    ) -> wgpu::TextureView {
-        let output = self.device.create_texture(&wgpu::TextureDescriptor {
-            label: None,
-            dimension: wgpu::TextureDimension::D2,
-            size: color.texture().size(),
-            format: wgpu::TextureFormat::Rgba8Unorm,
-            usage: wgpu::TextureUsages::TEXTURE_BINDING | wgpu::TextureUsages::STORAGE_BINDING,
-            view_formats: &[],
-            mip_level_count: 1,
-            sample_count: 1,
-        });
-
-        let view = output.create_view(&Default::default());
-        self.converter
-            .convert_color(&self.device, format, color, &view, encoder);
-
-        view
-    }
-
-    fn encode_depth(
-        &mut self,
-        depth: &wgpu::TextureView,
-        format: DepthCopyFormat,
-        encoder: &mut wgpu::CommandEncoder,
-    ) -> wgpu::TextureView {
-        let output = self.device.create_texture(&wgpu::TextureDescriptor {
-            label: None,
-            dimension: wgpu::TextureDimension::D2,
-            size: depth.texture().size(),
-            format: wgpu::TextureFormat::Rgba8Unorm,
-            usage: wgpu::TextureUsages::TEXTURE_BINDING | wgpu::TextureUsages::STORAGE_BINDING,
-            view_formats: &[],
-            mip_level_count: 1,
-            sample_count: 1,
-        });
-
-        let view = output.create_view(&Default::default());
-        self.converter
-            .convert_depth(&self.device, format, depth, &view, encoder);
-
-        view
-    }
-
-    fn get_texture_data(
-        &mut self,
-        view: &wgpu::TextureView,
-        mut encoder: wgpu::CommandEncoder,
-    ) -> Vec<u32> {
-        let size = view.texture().size();
-        let row_size = size.width * 4;
-        let row_stride = row_size.next_multiple_of(256);
-
-        encoder.copy_texture_to_buffer(
-            wgpu::TexelCopyTextureInfo {
-                texture: view.texture(),
-                mip_level: 0,
-                origin: wgpu::Origin3d::ZERO,
-                aspect: wgpu::TextureAspect::default(),
-            },
-            wgpu::TexelCopyBufferInfo {
-                buffer: &self.data_read_buffer,
-                layout: wgpu::TexelCopyBufferLayout {
-                    offset: 0,
-                    bytes_per_row: Some(row_stride),
-                    rows_per_image: None,
-                },
-            },
-            size,
-        );
-
-        let (sender, receiver) = oneshot::channel();
-        encoder.map_buffer_on_submit(&self.data_read_buffer, wgpu::MapMode::Read, .., |r| {
-            sender.send(r).unwrap()
-        });
-
-        let cmd = encoder.finish();
-        let submission = self.queue.submit([cmd]);
-        self.device
-            .poll(wgpu::wgt::PollType::Wait {
-                submission_index: Some(submission),
-                timeout: None,
-            })
-            .unwrap();
-
-        let result = receiver.recv().unwrap();
-        result.unwrap();
-
-        let mapped = self.data_read_buffer.get_mapped_range(..);
-        let data = &*mapped;
-
-        let mut texels = Vec::with_capacity(size.width as usize * size.height as usize);
-        for row in 0..size.height as usize {
-            let row_data = &data[row * row_stride as usize..][..row_size as usize];
-            texels.extend(
-                row_data
-                    .chunks_exact(4)
-                    .map(u32::read_from_bytes)
-                    .map(Result::unwrap),
-            );
-        }
-
-        std::mem::drop(mapped);
-        self.data_read_buffer.unmap();
-
-        texels
-    }
-
-    fn clear(&mut self, x: u32, y: u32, width: u32, height: u32) {
-        let color = self
-            .pipeline_settings
-            .blend
-            .color_write
-            .then_some(self.clear_color);
-        let depth = self
-            .pipeline_settings
-            .depth
-            .write
-            .then_some(self.clear_depth);
-
-        self.current_pass.set_scissor_rect(x, y, width, height);
-        self.current_pass
-            .set_viewport(0.0, 0.0, 640.0, 528.0, 0.0, 1.0);
-        self.cleaner
-            .clear_target(color, depth, &mut self.current_pass);
-    }
-
-    fn copy_color(
-        &mut self,
-        args: CopyArgs,
-        format: ColorCopyFormat,
-        response: Option<Sender<Texels>>,
-        id: TextureId,
-    ) {
-        let CopyArgs {
-            src,
-            dims,
-            half,
-            clear,
-        } = args;
-
-        self.debug(format!(
-            "color copy requested: ({}, {}) [{}x{}] (mip: {})",
-            src.x().value(),
-            src.y().value(),
-            dims.width(),
-            dims.height(),
-            half
-        ));
-
-        self.submit();
-
-        let mut encoder = self
-            .device
-            .create_command_encoder(&wgpu::CommandEncoderDescriptor::default());
-
-        let raw_texture = self.copy_color_to_tex(
-            src.x().value(),
-            src.y().value(),
-            dims.width(),
-            dims.height(),
-            half,
-            &mut encoder,
-        );
-
-        let encoded_texture = self.encode_color(&raw_texture, format, &mut encoder);
-        if let Some(response) = response {
-            let data = self.get_texture_data(&raw_texture, encoder);
-            response.send(data).unwrap();
-        } else {
-            let cmd = encoder.finish();
-            self.queue.submit([cmd]);
-        }
-
-        self.texture_cache.insert_direct(id, encoded_texture);
-        if clear {
-            self.clear(
-                src.x().value() as u32,
-                src.y().value() as u32,
-                dims.width() as u32,
-                dims.height() as u32,
-            );
-        }
-    }
-
-    fn copy_depth(
-        &mut self,
-        args: CopyArgs,
-        format: DepthCopyFormat,
-        response: Option<Sender<Texels>>,
-        id: TextureId,
-    ) {
-        let CopyArgs {
-            src,
-            dims,
-            half,
-            clear,
-        } = args;
-
-        self.debug(format!(
-            "depth copy requested: ({}, {}) [{}x{}] (mip: {})",
-            src.x().value(),
-            src.y().value(),
-            dims.width(),
-            dims.height(),
-            half
-        ));
-
-        self.submit();
-
-        let mut encoder = self
-            .device
-            .create_command_encoder(&wgpu::CommandEncoderDescriptor::default());
-
-        let raw_texture = self.copy_depth_to_tex(
-            src.x().value(),
-            src.y().value(),
-            dims.width(),
-            dims.height(),
-            half,
-            &mut encoder,
-        );
-
-        let encoded_texture = self.encode_depth(&raw_texture, format, &mut encoder);
-        if let Some(response) = response {
-            let data = self.get_texture_data(&raw_texture, encoder);
-            response.send(data).unwrap();
-        } else {
-            let cmd = encoder.finish();
-            self.queue.submit([cmd]);
-        }
-
-        self.texture_cache.insert_direct(id, encoded_texture);
-        if clear {
-            self.clear(
-                src.x().value() as u32,
-                src.y().value() as u32,
-                dims.width() as u32,
-                dims.height() as u32,
-            );
-        }
-    }
-
-    fn copy_xfb(&mut self, args: CopyArgs, id: u32) {
-        let CopyArgs {
-            src,
-            dims,
-            half,
-            clear,
-        } = args;
-
-        assert!(!half);
-
-        self.debug("XFB copy requested");
-        self.submit();
-
-        let x = src.x().value() as u32;
-        let y = src.y().value() as u32;
-        let width = dims.width() as u32;
-        let height = dims.height() as u32;
-
-        let size = wgpu::Extent3d {
-            width,
-            height,
-            depth_or_array_layers: 1,
-        };
-
-        let color = self.embedded_fb.color();
-        let target = self.external_fb.create_copy(&self.device, id, size);
-
-        self.current_transfer_encoder.copy_texture_to_texture(
-            wgpu::TexelCopyTextureInfo {
-                texture: color.texture(),
-                mip_level: 0,
-                origin: wgpu::Origin3d { x, y, z: 0 },
-                aspect: wgpu::TextureAspect::default(),
-            },
-            wgpu::TexelCopyTextureInfo {
-                texture: target.texture(),
-                mip_level: 0,
-                origin: wgpu::Origin3d::ZERO,
-                aspect: wgpu::TextureAspect::default(),
-            },
-            size,
-        );
-
-        if clear {
-            self.clear(x, y, width, height);
-        }
-    }
-
-    fn present_xfb(&mut self, parts: Vec<XfbPart>) {
-        self.external_fb
-            .build(&mut self.current_transfer_encoder, parts);
-
-        self.submit();
     }
 }
